Move plugins and theme and resolve dependencies
