+++ docs/learn/glossary.mdx
@@ -3,6 +3,6 @@ title: Glossary
 description: Glossary of all specialized names and phrases used in the IOTA and Shimmer networks. If you dont understand a wording you will find answers here.
 ---
 
-import Glossary from '/theme/src/theme/Glossary';
+import Glossary from '@theme/Glossary';
 
 <Glossary />
