+++ package.json
@@ -16,7 +16,7 @@
     "format": "prettier --cache --write .",
     "format:check": "prettier --cache --check .",
     "format:svg": "svgo ./src/**/*.svg",
-    "prepare": "yarn build:plugins && yarn build:theme && yarn checkout",
+    "prepare": "yarn checkout",
     "docusaurus": "docusaurus",
     "start": "NODE_OPTIONS='--max-old-space-size=4096' docusaurus start",
     "start:section:build": "SELECTED_SECTION='build' yarn start",
@@ -42,6 +42,7 @@
     "@iota-wiki/plugin-tutorial": "workspace:^",
     "@iota-wiki/theme": "workspace:^",
     "@mdx-js/react": "^1.6.21",
+    "@artsy/to-title-case": "^1.1.0",
     "@popperjs/core": "^2.11.5",
     "@saucelabs/theme-github-codeblock": "^0.2.3",
     "@svgr/webpack": "^5.5.0",
@@ -54,6 +55,7 @@
     "flickity-fade": "^2.0.0",
     "globby": "^13.1.4",
     "hast-util-is-element": "1.1.0",
+    "html-react-parser": "^5.1.10",
     "humanize-duration": "^3.30.0",
     "infima": "^0.2.0-alpha.43",
     "plugin-image-zoom": "flexanalytics/plugin-image-zoom",
@@ -75,6 +77,7 @@
     "remark-remove-comments": "^0.2.0",
     "url-loader": "^4.1.1",
     "utf-8-validate": "^6.0.3",
+    "web3": "^4.8.0",
     "webpack": "^5.89.0",
     "webpack-node-externals": "^3.0.0",
     "ws": "^8.16.0"
