--- plugins/docs/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2019",
-    "module": "NodeNext",
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    "rootDir": "src",
-    "outDir": "dist",
-    "declaration": true
-  },
-  "include": ["src"]
-}
