--- plugins/tutorial/.gitignore
@@ -1,10 +1,0 @@
-node_modules
-/dist
-
-.yarn/*
-!.yarn/patches
-!.yarn/plugins
-!.yarn/releases
-!.yarn/sdks
-!.yarn/versions
-.pnp.*
