+++ src/plugins/cookiebot/index.js
