+++ src/plugins/tutorial/index.ts
@@ -4,8 +4,23 @@ import type {
   Plugin,
 } from '@docusaurus/types';
 
-import type { NormalizedOptions, UserOptions } from './types';
+import type { NormalizedOptions, UserOptions } from '../typesypes';
 
+export interface Options {
+  title: string;
+  description: string;
+  tags: string[];
+  source?: string;
+  preview?: string;
+}
+
+export interface DefaultOptions {
+  id: string;
+  route: string;
+}
+
+export type UserOptions = Options & Partial<DefaultOptions>;
+export type NormalizedOptions = Options & DefaultOptions;
 export default async function pluginTutorial(
   context: LoadContext,
   options: NormalizedOptions,
