--- theme/README.md
@@ -1,72 +1,0 @@
-# IOTA Theme
-
-The IOTA Wiki theme for Docusaurus.
-
-## Installation
-
-Add `@iota-wiki/theme` to your package:
-
-```bash
-npm i @iota-wiki/theme
-# or
-yarn add @iota-wiki/theme
-```
-
-Modify your `docusaurus.config.js`:
-
-```diff
-module.exports = {
-  ...
-+ themes: ['@iota-wiki/theme'],
-  ...
-}
-```
-
-## Components
-
-### AddToMetaMaskButton
-
-A button to add the ShimmerEVM network to MetaMask.
-
-```
-import { AddToMetaMaskButton } from '@theme/AddToMetaMaskButton'
-import { Networks } from '@theme/constant'
-
-<AddToMetaMaskButton {...Networks['shimmer']}/>
-```
-
-### CodeSnippetBlock
-
-This component finds `startString` and `endString` in a code listing and displays the snippet in between.
-
-```
-import CodeSnippetBlock from '@theme/CodeSnippetBlock'
-import code_snippet from '!!raw-loader!../../code_snippet.py'
-
-<CodeSnippetBlock
-  language={'python'}
-  code={code_snippet}
-  startString={'# start'}
-  endString={'# end'} />
-```
-
-### NetworkInfo
-
-A set of components showing information about the different networks.
-
-```
-import NetworkInfo from '@theme/NetworkInfo'
-import { Networks } from '@theme/constant'
-
-<NetworkInfo.L1  {...Networks['iota']}/>
-<NetworkInfo.Evm  {...Networks['iota']}/>
-<NetworkInfo.EvmCustom  {...Networks['iota']}/>
-```
-
-## Swizzling components
-
-```bash
-npm swizzle @iota-wiki/theme [component name]
-```
-
-All components used by this theme can be found [here](https://github.com/iota-wiki/iota-wiki/tree/main/theme/src/theme)
