--- theme/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "@iota-wiki/theme",
-  "version": "0.1.0",
-  "description": "IOTA Wiki theme components",
-  "main": "dist/index.js",
-  "types": "src/types.d.ts",
-  "license": "MIT",
-  "scripts": {
-    "copy": "copyfiles -u 1 \"./src/**/*.css\" \"./dist\"",
-    "build": "tsc --build && yarn copy && prettier --config ../.prettierrc.json --write \"dist/theme/**/*.js\"",
-    "watch": "nodemon --watch src --ext '*' --exec yarn build",
-    "prepack": "yarn build",
-    "test": "jest"
-  },
-  "dependencies": {
-    "@artsy/to-title-case": "^1.1.0",
-    "@docusaurus/theme-classic": "2.4.3",
-    "@docusaurus/theme-common": "2.4.3",
-    "@iota-wiki/plugin-docs": "workspace:^",
-    "@metamask/providers": "^10.2.1",
-    "clsx": "^1.2.1",
-    "html-react-parser": "^4.2.10",
-    "react-markdown": "6",
-    "web3": "^4.2.2"
-  },
-  "devDependencies": {
-    "@docusaurus/types": "2.4.3",
-    "@types/react": "18.2.71",
-    "copyfiles": "^2.4.1",
-    "nodemon": "^2.0.16",
-    "prettier": "^2.8.8",
-    "react": "18.2.0",
-    "typescript": "^4.9"
-  },
-  "peerDependencies": {
-    "react": "*"
-  },
-  "files": [
-    "./src",
-    "./dist"
-  ],
-  "packageManager": "yarn@3.2.0",
-  "engines": {
-    "node": ">=14.0.0"
-  }
-}
