--- theme/tsconfig.server.json
@@ -1,14 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    /* Emit */
-    "rootDir": "src",
-    "outDir": "dist",
-    "noEmit": false,
-    "composite": true,
-    "incremental": true,
-    "tsBuildInfoFile": "./dist/.tsbuildinfo"
-  },
-  "include": ["src"],
-  "exclude": ["src/theme", "**/__tests__/**"]
-}
